requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --resolver=backtracking ./requirements.in
#
aiohttp==3.8.4
    # via openai
aiolimiter==1.0.0
    # via python-telegram-bot
aiosignal==1.3.1
    # via aiohttp
anyio==3.6.2
    # via httpcore
apscheduler==3.10.1
    # via python-telegram-bot
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
cachetools==5.3.0
    # via python-telegram-bot
certifi==2023.5.7
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
httpcore==0.17.0
    # via httpx
httpx==0.24.0
    # via python-telegram-bot
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
openai==0.27.6
    # via -r ./requirements.in
pyaes==1.6.1
    # via pyrogram
pyrogram==2.0.106
    # via -r ./requirements.in
pysocks==1.7.1
    # via pyrogram
python-dotenv==1.0.0
    # via -r ./requirements.in
python-telegram-bot[ext]==20.3
    # via -r ./requirements.in
pytz==2023.3
    # via
    #   apscheduler
    #   python-telegram-bot
requests==2.30.0
    # via openai
six==1.16.0
    # via apscheduler
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
tgcrypto==1.2.5
    # via -r ./requirements.in
tornado==6.3.2
    # via python-telegram-bot
tqdm==4.65.0
    # via openai
tzlocal==5.0.1
    # via apscheduler
urllib3==2.0.2
    # via requests
uvloop==0.17.0
    # via -r ./requirements.in
yarl==1.9.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
